requirements.txt
ccxt==1.91.93
cffi==1.15.0
cryptography==37.0.2
gevent==21.12.0
greenlet==1.1.2
ipython==8.4.0
jwt==1.3.1
pycparser==2.21
pydantic==1.9.1
websocket-client==1.3.2
zope.event==4.5.0
zope.interface==5.4.0
